+++ src/Articulate/Controllers/ArticulateAuthorsController.cs
@@ -2,8 +2,10 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ViewEngines;
 using Microsoft.Extensions.Logging;
+using Umbraco.Cms.Core.Cache;
 using Umbraco.Cms.Core.Models.PublishedContent;
 using Umbraco.Cms.Core.Web;
+using Umbraco.Cms.Web.Common;
 using Umbraco.Cms.Web.Common.Controllers;
 
 namespace Articulate.Controllers
@@ -12,17 +14,21 @@ namespace Articulate.Controllers
     /// Handles the ArticulateAuthors container node (/authors/).
     /// </summary>
     /// <remarks>
-    /// By default, redirects to blog root when redirectArchive is enabled.
     /// If a custom theme provides Authors.cshtml, it will render the author listing.
+    /// Otherwise, redirects to the blog root.
     /// Individual author pages (/authors/john-doe/) are handled by <see cref="ArticulateAuthorController"/>
     /// </remarks>
     public class ArticulateAuthorsController(
         ILogger<ArticulateAuthorsController> logger,
         ICompositeViewEngine compositeViewEngine,
         IUmbracoContextAccessor umbracoContextAccessor,
+        UmbracoHelper umbracoHelper,
+        AppCaches appCaches,
         IPublishedValueFallback publishedValueFallback)
         : RenderController(logger, compositeViewEngine, umbracoContextAccessor)
     {
+        private AppCaches AppCaches { get; } = appCaches;
+
         /// <inheritdoc/>
         public override IActionResult Index()
         {
@@ -34,39 +40,74 @@ namespace Articulate.Controllers
 
             var root = new MasterModel(CurrentPage, publishedValueFallback);
 
-            if (root.RootBlogNode.Value<bool>("redirectArchive"))
-            {
-                return RedirectPermanent(root.RootBlogNode.Url());
-            }
-
             // Check if theme has custom Authors.cshtml view before building the listing model.
             if (!EnsurePhysicalViewExists("Authors"))
             {
                 logger.LogInformation(
                     "ArticulateAuthorsController: No Authors.cshtml view found. " +
-                    "Recommend enabling 'redirectArchive' or creating custom Authors.cshtml in theme.");
+                    "Redirecting authors directory to blog root.");
 
-                return NotFound();
+                return RedirectPermanent(root.RootBlogNode.Url());
             }
 
-            // Build author list for custom themes that provide Authors.cshtml
-            var listingPager = new PagerModel(1, 0, 1);
+            IPublishedContent[] listNodes = root.RootBlogNode.Children()
+                .Where(x => x.ContentType.Alias == ArticulateConstants.ContentType.ArticulateArchive)
+                .ToArray();
+            int[] listNodeIds = listNodes.Select(x => x.Id).ToArray();
+            IReadOnlyDictionary<string, (int PostCount, DateTime? LastPostDate)> authorStats =
+                GetAuthorPostStats(root.RootBlogNode.Id, listNodeIds);
+
             var authorNodes = CurrentPage.Children().ToList();
             var authors = authorNodes
-                .Select(a => new AuthorModel(
-                    a,
-                    [], // Author listing pages do not need post items.
-                    listingPager,
-                    0,
-                    publishedValueFallback))
+                .Select(a =>
+                {
+                    var authorKey = UmbracoHelperExtensions.NormalizeAuthorName(a.Name);
+                    _ = authorStats.TryGetValue(authorKey, out (int PostCount, DateTime? LastPostDate) stats);
+
+                    Umbraco.Cms.Core.Models.MediaWithCrops? image = a.Value<Umbraco.Cms.Core.Models.MediaWithCrops>("authorImage");
+
+                    return new AuthorDirectoryItemModel
+                    {
+                        Name = a.Name,
+                        Bio = a.Value<string>("authorBio") ?? string.Empty,
+                        AuthorUrl = a.Value<string>("authorUrl").ToSafeHrefUrl(),
+                        BlogUrl = a.Url(),
+                        AuthorRssUrl = root.RootBlogNode.Url(mode: UrlMode.Absolute).EnsureEndsWith('/') + "author/" + a.Id + "/rss",
+                        Image = image,
+                        CroppedWideUrl = image?.GetCropUrl(cropAlias: "wide", preferFocalPoint: true, useCropDimensions: true) ?? string.Empty,
+                        PostCount = stats.PostCount,
+                        LastPostDate = stats.LastPostDate
+                    };
+                })
                 .ToList();
 
-            var model = new AuthorListModel(CurrentPage, publishedValueFallback)
+            var model = new AuthorDirectoryModel(CurrentPage, publishedValueFallback)
             {
                 Authors = authors
             };
 
             return View("Authors", model);
+        }
+
+        private IReadOnlyDictionary<string, (int PostCount, DateTime? LastPostDate)> GetAuthorPostStats(
+            int rootBlogNodeId,
+            int[] listNodeIds)
+        {
+            IReadOnlyDictionary<string, (int PostCount, DateTime? LastPostDate)> GetResult() =>
+                umbracoHelper.GetAuthorPostStatsByAuthor(listNodeIds);
+
+#if DEBUG
+            return GetResult();
+#else
+            return (IReadOnlyDictionary<string, (int PostCount, DateTime? LastPostDate)>)AppCaches.RuntimeCache.Get(
+                string.Concat(
+                    nameof(ArticulateAuthorsController),
+                    nameof(UmbracoHelperExtensions.GetAuthorPostStatsByAuthor),
+                    rootBlogNodeId,
+                    string.Join(",", listNodeIds.Order())),
+                GetResult,
+                TimeSpan.FromSeconds(30))!;
+#endif
         }
     }
 }